Move CommitmentTreeWitness into zcash_primitives
